bot settings: extract ShipPickerSheet into shared ui components
Move the ship picker sheet out of BotShipListSettingsScreen so other
screens can reuse it. Title and disabled-reason copy become optional
props with the previous strings as defaults.
